extract Tooltips
